+++ packages/board/src/useScanBoard.ts
@@ -293,6 +293,11 @@ export function useScanBoard(options: UseScanBoardOptions | string = {}) {
     [modeler],
   );
 
+  const duplicateBoundary = useCallback(
+    (id: string) => modeler.modeling.duplicateBoundary(id),
+    [modeler],
+  );
+
   const deleteConnection = useCallback(
     (id: string) => modeler.modeling.deleteConnection(id),
     [modeler],
@@ -496,6 +501,7 @@ export function useScanBoard(options: UseScanBoardOptions | string = {}) {
       redo,
       deleteElement,
       duplicateElement,
+      duplicateBoundary,
       deleteConnection,
       updateConnection,
       renameElement,
@@ -543,6 +549,7 @@ export function useScanBoard(options: UseScanBoardOptions | string = {}) {
       redo,
       deleteElement,
       duplicateElement,
+      duplicateBoundary,
       deleteConnection,
       updateConnection,
       renameElement,